Avoid recomputing static error messages on subsequent parser runs.
For parsers like [match], [match1], [skip_while] and [take_while], the
error messages they produce are only a function of the arguments used to
construct them rather than any of the input. This change makes it such
that we lazily compute and store such error messages for future use.

Performance improvements are reflected in the JSON benchmarks.